deletion
This lambda function will check for "DeleteOn" tags which is attached with snapshots as created by creation function. This snapshots contains a DeleteOn tag with retention days added. If the date matches with current date then the snapshots will be deleted. This can run on a daily basis , so that all the snapshots with correct retention date will be deleted.